+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 - strip and lower username for register user
 - clear database for sqllite
+- datetime.time as api parameter data type
 
 ### 2018 01 21